fixed: ArrayIndexOutOfBoundsException when reading test input > 32KB

+++ pact/pact-tests/src/test/java/eu/stratosphere/pact/test/testPrograms/util/IntTupleDataInFormat.java
@@ -26,7 +26,7 @@ public class IntTupleDataInFormat extends DelimitedInputFormat {
 	public static final int DELIMITER = '|';
 	
 	private final PactInteger key = new PactInteger();
-	private final short[] offsets = new short[MAX_COLUMNS];
+	private final int[] offsets = new int[MAX_COLUMNS];
 
 	@Override
 	public boolean readRecord(PactRecord target, byte[] line, int offset, int numBytes)
@@ -35,14 +35,14 @@ public class IntTupleDataInFormat extends DelimitedInputFormat {
 		int readPos = offset;
 
 		// allocate the offsets array
-		final short[] offsets = this.offsets;
-		offsets[0] = (short) offset;
+		final int[] offsets = this.offsets;
+		offsets[0] = offset;
 
 		int col = 1; // the column we are in
 
 		while (readPos < limit) {
 			if (line[readPos++] == DELIMITER) {
-				offsets[col++] = (short) (readPos);
+				offsets[col++] = readPos;
 			}
 		}
